+++ src/frontend/apps/impress/src/hook/useCmdK.tsx
@@ -5,6 +5,13 @@ export const useCmdK = (callback: () => void) => {
     const down = (e: KeyboardEvent) => {
       if ((e.key === 'k' || e.key === 'K') && (e.metaKey || e.ctrlKey)) {
         e.preventDefault();
+
+        const isProseMirrorFocused =
+          document.activeElement?.classList.contains('ProseMirror');
+        if (isProseMirrorFocused) {
+          return;
+        }
+
         callback();
       }
     };